+++ app/playlists_ui/playlistscontroller.cpp
@@ -186,12 +186,19 @@ namespace PlaylistsUi {
   }
 
   void Controller::on_search(const QString &term) {
-    proxy->setFilterWildcard("*" + term + "*");
+#if (QT_VERSION >= QT_VERSION_CHECK(6, 0, 0))
+    QRegularExpression regex(term);
+    proxy->setFilterRegularExpression(regex);
+#else
+    QRegExp regex(term, Qt::CaseInsensitive, QRegExp::Wildcard);
+    proxy->setFilterRegExp(regex);
+#endif
+    proxy->setFilterCaseSensitivity(Qt::CaseInsensitive);
 
-    QTimer::singleShot(20, [=]() {
+    /*QTimer::singleShot(20, [=]() {
       if (!view->selectionModel()->selectedRows().isEmpty()) {
         view->scrollTo(view->selectionModel()->selectedRows().first(), QAbstractItemView::PositionAtCenter);
       }
-    });
+    });*/
   }
 }