+++ win/installer_win64_qt6_host.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-QT_PATH=/e/Qt/6.6.2/mingw_64/
+QT_PATH=/e/Qt/6.7.1/mingw_64/
 TOOLCHAIN_PATH=/e/Qt/Tools/mingw1120_64/bin
-INSTALLER_FRAMEWORK_PATH=/e/Qt/Tools/QtInstallerFramework/4.6/bin
+INSTALLER_FRAMEWORK_PATH=/e/Qt/Tools/QtInstallerFramework/4.8/bin
 
 export PATH=$PATH:$TOOLCHAIN_PATH:$QT_PATH/bin/:$INSTALLER_FRAMEWORK_PATH